revert to last working attempt 2

+++ COMP30019 Project 1/Assets/Scripts/Terrain Generation/HeightGrid.cs
@@ -16,7 +16,7 @@ public class HeightGrid
 		this.grid = new Vector3[size * size];
 		this.size = size;
 	}
-	public float GetGridSize()
+	public int GetGridSize()
 	{
 		return size;
 	}
